2020-02-14, Version 12.16.1 'Erbium' (LTS)
Notable changes:

Node.js 12.16.0 included 3 regressions that are being fixed in this release

Accidental Unflagging of Self Resolving Modules:

12.16.0 included a large update to the ESM implementation. One of
the new features, Self Referential Modules, was accidentally released
without requiring the `--experimental-modules` flag. This release is
being made to appropriately flag the feature.

Process Cleanup Changed Introduced WASM-Related Assertion

A change during Node.js process cleanup led to a crash in combination with
specific usage of WASM. This has been fixed by partially reverted said change.
A regression test and a full fix are being worked on and will likely be included
in future 12.x and 13.x releases.

Use Largepages Runtime Option Introduced Linking Failure:

A Semver-Minor change to introduce `--use-largepages` as a runtime
option introduced a linking failure. This had been fixed in master
but regressed as the fix has not yet gone out in a Current release.
The feature has been reverted, but will be able to re-land with a fix
in a future Semver-Minor release.

Async Hooks was Causing an Exception When Handling Errors:

Changes in async hooks internals introduced a case where an internal
api call could be called with undefined causing a process to crash.
The change to async hooks was reverted. A regression test and fix has
been proposed and the change could re land in a future Semver-Patch
release if the regression is reliably fixed.

PR-URL: https://github.com/nodejs/node/pull/31781